Handle external auth problems
This commit adds a middleware which catches a custom exception
`ImproperCredentialsException` which should be raised by modules when
there is an authentication problem. If left uncaught, this exception
will be handled by the custom middleware which will redirect the user to
the homepage and display an error.

Fixes https://github.com/TOMToolkit/tom_base/issues/38