Make required fields for non-sid. target creation depend on scheme
Previously, *all* fields were required in the non-sidereal target
creation form, including fields which should be optional (e.g.`name2`,
`name3`) and fields that are not applicable to particular schemes (e.g.
`mean_daily_motion` is not needed for comets, but was required by the
validation).

This commit allows each scheme to have a list of required fields, and
allows optional/irrelevant fields to be left blank in the form.

Also fixes a minor bug where after submitting the non-sidereal form with
validation errors, the sidereal tab is selected but the non-sidereal
form is shown.