Stop infinite loop for hint preferences in tom_setup script
Also, the template settings.tmpl previously set `HINTS_ENABLED` to a
string, e.g.  `'True'` or `'False'`. The string `'False'` evaluates to
true when used as the expression in an if statement, so this resulted in
hints being shown even when the user answers 'N' in `tom_setup`.

This commit fixes the template so that it is a boolean.