Adds support for Lasair's next gen broker Iris
The LasairIris module works by querying a "Stream Log" stored on Lasair
Iris. A steam log is a log from a stored query on lasair that is updated
in near real time and written to a static file, so no sql needs to be
run when the user requests it.

I need to support this in a TOM I'm building for another project so I
figured I'd send it upstream.