Make TOM name configurable
It would be nice if users would not have to override
tom_common/base.html just to update the name of the TOM that is
displayed in the header and title of each page. I think those templates
are just fine.

This introduces a backwards compatible tag that pulls a TOM_NAME setting
from settings.py but defaults to "TOM Toolkit" in case that setting
doesn't exist. I added the setting to the setup template, but it might
be a good candidate to add to the setup script as an interactive
question as well.