Seperate Target list table into a partial
It would be helpful to have the target list as it's own template so that
we can override it without overriding the entire page. A good example
would be adding columns to display in the table, while still being able
to get updates from upstream on the rest of the page.