Adds a move function for the target list
This PR adds functions for moving targets between TargetLists. If a
target is moved to a TargetList, it's previous lists are cleared and
the target is added to the specified TargetList. If the target already
exists in the specified TargetList a warning is displayed but the
target's previous associations are not changed.

In addition a button is added to the top of the target table alongside
the add/remove buttons.

This is a faster way of doing remove/add for a large quantity of
targets.