Adds an ordering option to the target list page
This PR adds a django_fitlers.OrderingFilter to the filter form on the
target list page, allowing users to choose the order in which to display
targets in the table.

The SD Tom requires this because we frequently update our targets and
would like targets with recent updates to appear higher in the list.