Adds a param to the airmass plot to render on load
This PR adds a parameter to the `target_plan` inclusion tag which allows
the plot to render on page load with default parameters equal to the
next 24 hours and 2.5 airmass.

The default behvior or requiring the form to filled out and submitted
remains unchanged.

This doesn't seem to have much of an effect on page load time so it
might be better to make this behavior the default. The param can even be
preserved but setting it causes the plot NOT to render by default.