Hides large TAGS header on target detail
On the target detail, there is a large, unsightly TAGS header item in
the left column, which appears whether or not the target has any tags
defined. This change hides the section unless the target has any tags,
as well as tones down the font size a bit even when it does.