Add index on BaseTarget.created to speed up load time of large DBs
We currently sort on the created field of BaseTarget, which causes
performance issues when there are a large number of BaseTargets.

In a database with 6 million rows, this reduced the time for the SQL
count(*) from ~800ms to ~100ms, significantly speeding up page load
time.

Note that this does NOT solve #598 but it does help a little.