Update target create/update forms for permissions
It doesn't make a lot of sense to display the group form on the target
create/update forms if the target is not private. This commit introduces
a UI using a small amount of javascript that will hide the group form
unless the user sets the target to private.

Because this wasn't possible to implement with the generic form
rendering in use, this commit also updates the target update
and create forms to use a more compact layout instead of
placing every field on a single row, increasing usability.
The fields will still collapse to single rows on
narrow viewports.

Custom target types will default to the generic form rendering as
before.