Visual representation of target groups
Adds representation of the groups a target belongs to in various parts
of the UI. Groups can be assigned a color upon creation.

On the target list, the groups appear as pills beneath the target name.
Clicking on a group pill will filter the table by that target group.

On the target detail page, clicking a group pill focuses the "manage
groups" tab.