Update UserSession after password change and keep user logged in
When the user changes their password good practice is to call
update_session_auth_hash to keep them logged in. This create a new
Session object. That means that the old UserSession object gets
deleted on CASCADE. This create a new UserSession, linking the
User with their new Session.

Previously, the call to update_session_auth_hash was in dispatch(),
but now it's here in form_valid().