save Fernet key to Session before Session is copied
update_session_auth_hash also copies the old Session object to a new
one. So, in addition to creating a new UserSession, we need to copy
the Fernet hash in to the new session (in the request) so it can be
included when update_session_auth_hash does it's thing. The Fernet
hash that we need has been stashed temporarily in the User object
by session_utils.reencrypt_sensitive_data().