Clarify Session vs SessionStore access to encryption key
The SessionStore is the 'normal' way to access the data in the
Session model via the dictionary-like API it provides. And this
is how we do it when we have a HTTPRequest instance that has been
decorated with a SessionStore object in the middleware. However,
when we need access without a decorated HTTPRequest, we use the
UserSession which gives us access (it has a ForeignKey) to the
underlying Session model. Then we use session.get_decode().